requirements.txt
# Core dependencies
tensorflow>=2.15,<2.16
tfx==1.15.1
kfp==2.0.0
apache-beam[gcp]==2.48.0

# Google Cloud client libraries
google-cloud-storage==2.10.0
google-cloud-bigquery>=3,<4
google-cloud-aiplatform==1.28.0
google-cloud-pubsub==2.17.0
google-cloud-datastore==2.17.0
google-cloud-logging==3.4.0
google-cloud-monitoring==2.14.1

# Data processing and visualization
pandas
scikit-learn
matplotlib
seaborn
plotly

# Utilities
python-dotenv
prometheus_client
psutil
gsutil
PyYAML

# Kubeflow Katib
kubeflow-katib==0.13.0

# Testing
pytest
pytest-cov

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.24.0

# API
fastapi==0.100.0
uvicorn==0.23.1
aiohttp

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.2

# Linting and formatting
black==23.7.0
flake8==6.0.0
isort==5.12.0
